refactoring: extract test helpers into test/helpers.js removing code duplication
