+++ truffle/test/foreign.js
@@ -1,4 +1,5 @@
 var ForeignBridge = artifacts.require("ForeignBridge");
+var helpers = require("./helpers/helpers");
 
 contract('ForeignBridge', function(accounts) {
   it("should deploy contract", function() {
@@ -290,29 +291,6 @@ contract('ForeignBridge', function(accounts) {
     })
   })
 
-  function sign(address, data) {
-    return new Promise(function(resolve, reject) {
-      web3.eth.sign(address, data, function(err, result) {
-        if (err !== null) {
-          return reject(err);
-        } else {
-          return resolve(normalizeSignature(result));
-          //return resolve(result);
-        }
-      })
-    })
-  }
-
-  // geth && testrpc has different output of eth_sign than parity
-  // https://github.com/ethereumjs/testrpc/issues/243#issuecomment-326750236
-  function normalizeSignature(signature) {
-    // strip 0x
-    signature = signature.substr(2);
-
-    // increase v by 27...
-    return "0x" + signature.substr(0, 128) + (parseInt(signature.substr(128), 16) + 27).toString(16);
-  }
-
   it("should successfully submit signature and trigger CollectedSignatures event", function() {
     var meta;
     var signature;
@@ -322,7 +300,7 @@ contract('ForeignBridge', function(accounts) {
     var message = "0x111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111";
     return ForeignBridge.new(requiredSignatures, authorities, withdrawRelayGasCostPerAuthority).then(function(instance) {
       meta = instance;
-      return sign(authorities[0], message);
+      return helpers.sign(authorities[0], message);
     }).then(function(result) {
       signature = result;
       return meta.submitSignature(result, message, { from: authorities[0] });
@@ -348,7 +326,7 @@ contract('ForeignBridge', function(accounts) {
     var message = "0x111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111";
     return ForeignBridge.new(requiredSignatures, authorities, withdrawRelayGasCostPerAuthority).then(function(instance) {
       meta = instance;
-      return sign(authorities[0], message);
+      return helpers.sign(authorities[0], message);
     }).then(function(result) {
       return meta.submitSignature(result, message, { from: authorities[0] });
     }).then(function(result) {
@@ -368,10 +346,10 @@ contract('ForeignBridge', function(accounts) {
     return ForeignBridge.new(requiredSignatures, authorities, withdrawRelayGasCostPerAuthority).then(function(instance) {
       meta = instance;
       return Promise.all([
-        sign(authorities[0], message),
-        sign(authorities[1], message),
-        sign(authorities[0], message2),
-        sign(authorities[1], message2),
+        helpers.sign(authorities[0], message),
+        helpers.sign(authorities[1], message),
+        helpers.sign(authorities[0], message2),
+        helpers.sign(authorities[1], message2),
       ]);
     }).then(function(result) {
       signatures_for_message.push(result[0]);
@@ -423,7 +401,7 @@ contract('ForeignBridge', function(accounts) {
     var message = "0x1111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111";
     return ForeignBridge.new(requiredSignatures, authorities, withdrawRelayGasCostPerAuthority).then(function(instance) {
       meta = instance;
-      return sign(authorities[0], message);
+      return helpers.sign(authorities[0], message);
     }).then(function(result) {
       return meta.submitSignature(result, message, { from: authorities[0] });
     }).then(function(result) {
@@ -442,7 +420,7 @@ contract('ForeignBridge', function(accounts) {
     var message2 = "0x111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111112";
     return ForeignBridge.new(requiredSignatures, authorities, withdrawRelayGasCostPerAuthority).then(function(instance) {
       meta = instance;
-      return sign(authorities[0], message);
+      return helpers.sign(authorities[0], message);
     }).then(function(result) {
       return meta.submitSignature(result, message2, { from: authorities[0] });
     }).then(function(result) {
@@ -460,7 +438,7 @@ contract('ForeignBridge', function(accounts) {
     var message = "0x111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111";
     return ForeignBridge.new(requiredSignatures, authorities, withdrawRelayGasCostPerAuthority).then(function(instance) {
       meta = instance;
-      return sign(authorities[0], message);
+      return helpers.sign(authorities[0], message);
     }).then(function(result) {
       return meta.submitSignature(result, message, { from: authorities[1] });
     }).then(function(result) {
@@ -478,7 +456,7 @@ contract('ForeignBridge', function(accounts) {
     var message = "0x111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111";
     return ForeignBridge.new(requiredSignatures, authorities, withdrawRelayGasCostPerAuthority).then(function(instance) {
       meta = instance;
-      return sign(authorities[0], message);
+      return helpers.sign(authorities[0], message);
     }).then(function(result) {
       return meta.submitSignature(result, message, { from: authorities[0] });
     }).then(function(result) {
@@ -499,7 +477,7 @@ contract('ForeignBridge', function(accounts) {
 
     return ForeignBridge.new(requiredSignatures, authorities, withdrawRelayGasCostPerAuthority).then(function(instance) {
       foreignBridge = instance;
-      return sign(authorities[0], message);
+      return helpers.sign(authorities[0], message);
     }).then(function(result) {
       signature = result;
       return foreignBridge.submitSignature.estimateGas(result, message, { from: authorities[0] });
