+++ truffle/test/home.js
@@ -1,4 +1,5 @@
 var HomeBridge = artifacts.require("HomeBridge");
+var helpers = require("./helpers/helpers");
 
 contract('HomeBridge', function(accounts) {
   it("should deploy contract", function() {
@@ -65,74 +66,15 @@ contract('HomeBridge', function(accounts) {
     })
   })
 
-  function sign(address, data) {
-    return new Promise(function(resolve, reject) {
-      web3.eth.sign(address, data, function(err, result) {
-        if (err !== null) {
-          return reject(err);
-        } else {
-          return resolve(normalizeSignature(result));
-          //return resolve(result);
-        }
-      })
-    })
-  }
-
-  function signatureToVRS(signature) {
-    signature = strip0x(signature);
-    var v = parseInt(signature.substr(64 * 2), 16);
-    var r = "0x" + signature.substr(0, 32 * 2);
-    var s = "0x" + signature.substr(32 * 2, 32 * 2);
-    return {v: v, r: r, s: s};
-  }
-
-  // geth && testrpc has different output of eth_sign than parity
-  // https://github.com/ethereumjs/testrpc/issues/243#issuecomment-326750236
-  function normalizeSignature(signature) {
-    signature = strip0x(signature);
-
-    // increase v by 27...
-    return "0x" + signature.substr(0, 128) + (parseInt(signature.substr(128), 16) + 27).toString(16);
-  }
-
-  // strips leading "0x"
-  function strip0x(input) {
-    return input.substr(2);
-  }
-
-  function bigNumberToHexString(num) {
-    web3._extend.utils.isBigNumber(num);
-    var quotient = num;
-    var result = "";
-    while (quotient > 0) {
-      var remainderDec = quotient.mod(16).toNumber();
-      assert(remainderDec < 16);
-      var remainderHexDigit = remainderDec.toString(16);
-      assert.equal(remainderHexDigit.length, 1)
-      result = remainderHexDigit + result;
-      quotient = quotient.dividedToIntegerBy(16);
-    }
-    return "0x" + result;
-  }
-
-  function bigNumberToPaddedBytes32(num) {
-    web3._extend.utils.isBigNumber(num);
-    var result = strip0x(bigNumberToHexString(num));
-    while (result.length < 64) {
-      result = "0" + result;
-    }
-    return "0x" + result;
-  }
-
   function createMessage(recipient, value, transactionHash) {
     web3._extend.utils.isBigNumber(value);
-    recipient = strip0x(recipient);
+    recipient = helpers.strip0x(recipient);
     assert.equal(recipient.length, 20 * 2);
 
-    transactionHash = strip0x(transactionHash);
+    transactionHash = helpers.strip0x(transactionHash);
     assert.equal(transactionHash.length, 32 * 2);
 
-    var value = strip0x(bigNumberToPaddedBytes32(value));
+    var value = helpers.strip0x(helpers.bigNumberToPaddedBytes32(value));
     assert.equal(value.length, 64);
     var message = "0x" + recipient + value + transactionHash;
     var expectedMessageLength = (20 + 32 + 32) * 2 + 2;
@@ -158,10 +100,10 @@ contract('HomeBridge', function(accounts) {
       })
     }).then(function(result) {
       message = createMessage(recipient_account, value, result.tx);
-      return sign(authorities[0], message);
+      return helpers.sign(authorities[0], message);
     }).then(function(result) {
       signature = result;
-      var vrs = signatureToVRS(signature);
+      var vrs = helpers.signatureToVRS(signature);
 
       return homeBridge.withdraw.estimateGas(
         [vrs.v],
@@ -173,7 +115,7 @@ contract('HomeBridge', function(accounts) {
     }).then(function(result) {
       console.log("estimated gas cost of HomeBridge.withdraw =", result);
 
-      var vrs = signatureToVRS(signature);
+      var vrs = helpers.signatureToVRS(signature);
       return homeBridge.withdraw(
         [vrs.v],
         [vrs.r],
